+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Gitaly changelog
 
-UNRELEASED
+v0.24.1
 
 - Return an empty array on field `ParentIds` of `GitCommit`s if it has none
   https://gitlab.com/gitlab-org/gitaly/merge_requests/237